flask-server/requirements.txt
Flask==2.3.2
pandas==2.0.3
numpy==1.25.0
joblib==1.3.1
scikit-learn==1.3.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
geopy==2.3.0
overpy==0.6
folium==0.14.0